requirements.txt
streamlit==1.38.0
ultralytics==8.3.0
torch==2.4.1
torchvision==0.19.1
pillow==10.4.0
plotly==5.24.1
numpy==2.1.1
opencv-python-headless==4.10.0.84
